requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file base.txt base.in
#

cmsplugin-filer==1.0.1
contextlib2==0.5.1        # via raven
django-admin-tools==0.7.2
django-appconf==1.0.1     # via cmsplugin-filer, django-compressor, django-systemjs
django-autoslug==1.9.3
django-axes==1.6.0
django-choices==1.4.2
django-classy-tags==0.7.2  # via django-cms, django-sekizai
django-cms==3.2.3
django-compressor==2.0
django-filer==1.1.1
django-flat-theme==1.1.3
django-formtools==1.0     # via django-cms
django-mptt==0.8.3        # via django-filer
django-polymorphic==0.8.1  # via django-filer
django-rosetta==0.7.11
django-sekizai==0.9.0     # via cmsplugin-filer, django-cms
django-sniplates==0.4.1
django-systemjs==1.2
django-treebeard==4.0     # via django-cms
Django==1.8.12            # via cmsplugin-filer, django-choices, django-classy-tags, django-cms, django-filer, django-formtools, django-mptt, django-rosetta, django-sniplates, django-treebeard, easy-thumbnails
djangocms-admin-style==1.1.0  # via django-cms
djangocms-text-ckeditor==2.9.1
easy-thumbnails==2.3      # via cmsplugin-filer, django-filer
html5lib==0.9999999       # via django-cms, djangocms-text-ckeditor
markdown==2.6.6
microsofttranslator==0.7  # via django-rosetta
Pillow==3.2.0             # via djangocms-text-ckeditor, easy-thumbnails
polib==1.0.7              # via django-rosetta
pytz==2016.3
raven==5.12.0
rcssmin==1.0.6            # via django-compressor
requests==2.9.1           # via django-rosetta, microsofttranslator
rjsmin==1.0.12            # via django-compressor
semantic-version==2.5.0   # via django-systemjs
six==1.10.0               # via django-appconf, django-choices, django-rosetta, html5lib, microsofttranslator
Unidecode==0.4.19         # via django-filer

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via django-polymorphic
